feat(onboarding): Updated Cubits for cleaner logic
