+++ lib/core/routes.dart
@@ -4,6 +4,7 @@ import 'package:fit_lovers/presentations/screens/log_in_screen.dart';
 import 'package:fit_lovers/presentations/screens/onboarding_screen.dart';
 import 'package:fit_lovers/presentations/screens/register_screen.dart';
 import 'package:fit_lovers/presentations/screens/welcome_view_screen.dart';
+
 import 'package:flutter/material.dart';
 
 // Routes always separate words by -
