+++ lib/domain/cubit/authentication/auth_state.dart
@@ -45,8 +45,3 @@ class AuthConsentUpdated extends AuthState {
   @override
   List<Object> get props => [consent];
 }
-
-class AuthNeedsOnboarding extends AuthState {
-  @override
-  List<Object?> get props => [];
-}
